refactor: align star/unstar/stars/whoami with ping-style command conventions
- Split the star command group into a star/ subdirectory with common.ts
  (shared performStarAction, helpers, StarOptions) plus one file per
  command (star.ts, unstar.ts, stars.ts), matching the deprecation/
  layout.
- Each command file now exposes commandNames, handler, help,
  cliOptionsTypes, and rcOptionsTypes as top-level exports, consumed via
  'import * as <cmd>' in index.ts, matching ping/search/unpublish.
- whoami.ts moves to the same top-level export style; fetchWhoami helper
  is still exported for use by the star commands.
