+++ registry-access/commands/src/index.ts
@@ -3,8 +3,10 @@ import * as undeprecate from './deprecation/undeprecate.js'
 import * as distTag from './distTag.js'
 import * as ping from './ping.js'
 import * as search from './search.js'
-import { star, stars, unstar } from './star.js'
+import * as star from './star/star.js'
+import * as stars from './star/stars.js'
+import * as unstar from './star/unstar.js'
 import * as unpublish from './unpublish.js'
-import { whoami } from './whoami.js'
+import * as whoami from './whoami.js'
 
 export { deprecate, distTag, ping, search, star, stars, undeprecate, unpublish, unstar, whoami }
