+++ registry-access/commands/src/star/common.ts
@@ -1,21 +1,21 @@
-import { docsUrl } from '@pnpm/cli.utils'
 import { pickRegistryForPackage } from '@pnpm/config.pick-registry-for-package'
 import { PnpmError } from '@pnpm/error'
 import { createGetAuthHeaderByURI } from '@pnpm/network.auth-header'
 import { createFetchFromRegistry, type CreateFetchFromRegistryOptions } from '@pnpm/network.fetch'
 import npa from '@pnpm/npm-package-arg'
 import type { Registries, RegistryConfig } from '@pnpm/types'
-import { renderHelp } from 'render-help'
 
-import { rcOptionsTypes } from './common.js'
-import { fetchWhoami } from './whoami.js'
-
-export { rcOptionsTypes }
+import { rcOptionsTypes as commonRcOptionsTypes } from '../common.js'
+import { fetchWhoami } from '../whoami.js'
 
 export function cliOptionsTypes (): Record<string, unknown> {
   return {
-    ...rcOptionsTypes(),
+    ...commonRcOptionsTypes(),
   }
+}
+
+export function rcOptionsTypes (): Record<string, unknown> {
+  return commonRcOptionsTypes()
 }
 
 export interface StarOptions extends CreateFetchFromRegistryOptions {
@@ -23,57 +23,13 @@ export interface StarOptions extends CreateFetchFromRegistryOptions {
   registries?: Registries
 }
 
-export const star = {
-  cliOptionsTypes,
-  commandNames: ['star'],
-  handler: starHandler,
-  help: (): string => renderHelp({
-    description: 'Marks a package as a favorite.',
-    url: docsUrl('star'),
-    usages: ['pnpm star <package>'],
-  }),
-  rcOptionsTypes,
+interface PackumentWithStars {
+  _rev?: string
+  users?: Record<string, boolean>
+  [key: string]: unknown
 }
 
-export const unstar = {
-  cliOptionsTypes,
-  commandNames: ['unstar'],
-  handler: unstarHandler,
-  help: (): string => renderHelp({
-    description: 'Removes a package from your favorites.',
-    url: docsUrl('unstar'),
-    usages: ['pnpm unstar <package>'],
-  }),
-  rcOptionsTypes,
-}
-
-export const stars = {
-  cliOptionsTypes,
-  commandNames: ['stars'],
-  handler: starsHandler,
-  help: (): string => renderHelp({
-    description: 'Lists all packages starred by a specific user.',
-    url: docsUrl('stars'),
-    usages: ['pnpm stars [<user>]'],
-  }),
-  rcOptionsTypes,
-}
-
-async function starHandler (opts: StarOptions, params: string[]): Promise<void> {
-  if (params.length === 0) {
-    throw new PnpmError('STAR_PACKAGE_REQUIRED', 'Package name is required')
-  }
-  await performStarAction(opts, params[0], true)
-}
-
-async function unstarHandler (opts: StarOptions, params: string[]): Promise<void> {
-  if (params.length === 0) {
-    throw new PnpmError('UNSTAR_PACKAGE_REQUIRED', 'Package name is required')
-  }
-  await performStarAction(opts, params[0], false)
-}
-
-async function performStarAction (opts: StarOptions, packageName: string, star: boolean): Promise<void> {
+export async function performStarAction (opts: StarOptions, packageName: string, star: boolean): Promise<void> {
   const encodedName = parsePackageSpecName(packageName)
   const registryUrl = normalizeRegistryUrl(
     pickRegistryForPackage(opts.registries ?? { default: 'https://registry.npmjs.org/' }, packageName)
@@ -108,21 +64,14 @@ async function performStarAction (opts: StarOptions, packageName: string, star: 
 
   if (!response.ok) {
     if (response.status === 404 || response.status === 405 || response.status === 400 || response.status === 500) {
-      return performLegacyStarAction({ opts, packageName, encodedName, star, registryUrl, authHeader, fetchFromRegistry })
+      return performLegacyStarAction({ packageName, encodedName, star, registryUrl, authHeader, fetchFromRegistry })
     }
     const errorBody = await response.text()
     throw new PnpmError('REGISTRY_ERROR', `Failed to ${star ? 'star' : 'unstar'} package: ${response.status} ${response.statusText}. ${errorBody}`)
   }
 }
 
-interface PackumentWithStars {
-  _rev?: string
-  users?: Record<string, boolean>
-  [key: string]: unknown
-}
-
 interface LegacyStarActionArgs {
-  opts: StarOptions
   packageName: string
   encodedName: string
   star: boolean
@@ -173,57 +122,7 @@ async function performLegacyStarAction (args: LegacyStarActionArgs): Promise<voi
   }
 }
 
-async function starsHandler (opts: StarOptions, params: string[]): Promise<string> {
-  const registryUrl = normalizeRegistryUrl(opts.registries?.default ?? 'https://registry.npmjs.org/')
-  const fetchFromRegistry = createFetchFromRegistry(opts)
-  const authHeader = getAuthHeaderForRegistry(opts.configByUri, registryUrl)
-
-  let username = params[0]
-  if (!username) {
-    if (!authHeader) {
-      throw new PnpmError('STARS_UNAUTHORIZED', 'You must be logged in to list your starred packages')
-    }
-    username = await fetchWhoami(registryUrl, fetchFromRegistry, authHeader)
-  }
-
-  if (!params[0]) {
-    const starUrl = new URL('./-/user/v1/star', registryUrl).href
-    const response = await fetchFromRegistry(starUrl, {
-      authHeaderValue: authHeader,
-    })
-    if (response.ok) {
-      const starsData = await response.json() as string[] | Record<string, unknown>
-      if (Array.isArray(starsData)) return starsData.join('\n')
-      if (typeof starsData === 'object' && starsData !== null) {
-        return Object.keys(starsData).join('\n')
-      }
-    }
-  }
-
-  const starsUrl = new URL(`./-/user/${encodeURIComponent(username)}/stars`, registryUrl).href
-  let response = await fetchFromRegistry(starsUrl, {
-    authHeaderValue: authHeader,
-  })
-
-  if (!response.ok) {
-    const utilStarsUrl = new URL(`./-/util/user/${encodeURIComponent(username)}/stars`, registryUrl).href
-    response = await fetchFromRegistry(utilStarsUrl, {
-      authHeaderValue: authHeader,
-    })
-  }
-
-  if (!response.ok) {
-    if (response.status === 404) {
-      throw new PnpmError('USER_NOT_FOUND', `User "${username}" not found`)
-    }
-    throw new PnpmError('REGISTRY_ERROR', `Failed to fetch stars: ${response.status} ${response.statusText}`)
-  }
-
-  const starsData = await response.json() as Record<string, unknown>
-  return Object.keys(starsData).join('\n')
-}
-
-function parsePackageSpecName (packageName: string): string {
+export function parsePackageSpecName (packageName: string): string {
   let parsed
   try {
     parsed = npa(packageName)
@@ -236,7 +135,7 @@ function parsePackageSpecName (packageName: string): string {
   return parsed.escapedName
 }
 
-function getAuthHeaderForRegistry (
+export function getAuthHeaderForRegistry (
   configByUri: Record<string, RegistryConfig> | undefined,
   registryUrl: string
 ): string | undefined {
@@ -244,6 +143,6 @@ function getAuthHeaderForRegistry (
   return getAuthHeader(registryUrl)
 }
 
-function normalizeRegistryUrl (registryUrl: string): string {
+export function normalizeRegistryUrl (registryUrl: string): string {
   return registryUrl.endsWith('/') ? registryUrl : `${registryUrl}/`
 }
